+++ shaders/lib/depth_of_field.glsl
@@ -1,4 +1,9 @@
+#include "/utils/depth.glsl"
+
+
+
 vec3 getBlurredColor(vec2 coord, float size  ARGS_OUT) {
+	#include "/import/pixelSize.glsl"
 	vec3 colorTotal = vec3(0.0);
 	colorTotal += texture2D(MAIN_BUFFER, coord + vec2(-1, -2) * pixelSize * size, 0).rgb * 0.574;
 	colorTotal += texture2D(MAIN_BUFFER, coord + vec2( 0, -2) * pixelSize * size, 0).rgb * 0.641;
@@ -26,12 +31,20 @@ vec3 getBlurredColor(vec2 coord, float size  ARGS_OUT) {
 
 
 
-void doDOF(inout vec3 color DEBUG_ARG_OUT  ARGS_OUT) {
+void doDOF(inout vec3 color  DEBUG_ARGS_OUT  ARGS_OUT) {
 	
 	#if DOF_LOCKED_FOCAL_PLANE == 1
+		#include "/import/invFar.glsl"
 		float focusDepth = DOF_FOCAL_PLANE_DISTANCE * invFar;
 	#else
-		float focusDepth = centerLinearDepthSmooth;
+		#ifdef IS_IRIS
+			#include "/import/viewSize.glsl"
+			float depth = texelFetch(DEPTH_BUFFER_ALL, ivec2(viewSize) / 2, 0).r;
+			float focusDepth = toLinearDepth(depth  ARGS_IN);
+		#else
+			#include "/import/centerLinearDepthSmooth.glsl"
+			float focusDepth = centerLinearDepthSmooth;
+		#endif
 	#endif
 	
 	float linearDepth = toLinearDepth(texelFetch(DEPTH_BUFFER_ALL, texelcoord, 0).r  ARGS_IN);
@@ -49,6 +62,7 @@ void doDOF(inout vec3 color DEBUG_ARG_OUT  ARGS_OUT) {
 		debugOutput = vec3(nearBlurAmount, farBlurAmount, 0.0);
 	#endif
 	
+	#include "/import/far.glsl"
 	float blurSizeMult = 1.0 / (linearDepth * far * 0.01 + 1.0);
 	vec3 nearBlur = getBlurredColor(texcoord, nearBlurAmount * blurSizeMult * DOF_NEAR_BLUR_SIZE  ARGS_IN);
 	vec3 farBlur = getBlurredColor(texcoord, farBlurAmount * blurSizeMult * DOF_FAR_BLUR_SIZE  ARGS_IN);